Check if the autoprovisioned JDK can satisfy toolchain spec early
If the autoprovisioned JDK is already downloaded but not suitable the
build was unpacking the downloaded archive again and using the result
violating the spec. This happened because the presence of the
downloaded archive was short-circuiting the spec check.

This CL fix both unneeded unpacking and following use of the wrong JDK
by checking if JDK for the spec can be downloaded first, before checking
for the downloaded archive's presence.